refactor: rename BaseOpenApiSpec to AbstractOpenApiSpec
