+++ lib/classes/AbstractOpenApiSpec.js
