+++ lib/classes/OpenApi2Spec.js
@@ -1,12 +1,12 @@
 const utils = require('../utils');
-const OpenApiSpec = require('./BaseOpenApiSpec');
+const AbstractOpenApiSpec = require('./AbstractOpenApiSpec');
 
 const doesOpenApiPathMatchPathname = (openApiPath, pathname) => {
   const pathInColonForm = utils.convertOpenApiPathToColonForm(openApiPath);
   return utils.doesColonPathMatchPathname(pathInColonForm, pathname);
 };
 
-class OpenApi2Spec extends OpenApiSpec {
+class OpenApi2Spec extends AbstractOpenApiSpec {
   findOpenApiPathMatchingPathname(pathname) {
     const openApiPath = this.paths().find(openApiPath =>
       doesOpenApiPathMatchPathname(openApiPath, pathname)
