+++ lib/classes/OpenApi3Spec.js
@@ -2,7 +2,7 @@ const { inspect } = require('util');
 const url = require('url');
 
 const utils = require('../utils');
-const OpenApiSpec = require('./BaseOpenApiSpec');
+const AbstractOpenApiSpec = require('./AbstractOpenApiSpec');
 
 const serversPropertyNotProvidedOrIsEmptyArray = (spec) =>
   (!spec.hasOwnProperty('servers') || !spec.servers.length);
@@ -23,7 +23,7 @@ const doesOpenApiPathMatchPathname = (openApiPath, pathname, matchingServerBaseP
   return openApiPathMatchesPathname;
 };
 
-class OpenApi3Spec extends OpenApiSpec {
+class OpenApi3Spec extends AbstractOpenApiSpec {
   constructor(spec) {
     super(spec);
     this.ensureDefaultServer();
